- Refactor CalikoDemo2D into using a factory CalikoDemoStructure2DFactory
